More fixes to format specifiers, mainly around use of NSUInteger.

+++ GSWeb/GSWRequest.m
@@ -608,7 +608,7 @@ RCS_ID("$Id$")
 //--------------------------------------------------------------------
 -(NSString*)description
 {
-  return [NSString stringWithFormat:@"<%s %p - method=%@, uri=%@, httpVersion=%@, headers=%@, content=%@, userInfo=%@, defaultFormValueEncoding=%u, formValueEncoding=%u, formValues=%@, uriElements=%@, cookie=%@, applicationURLPrefix=%@, requestHandlerPathArray=%@, browserLanguages=%@, requestType=%d, isUsingWebServer=%s, formValueEncodingDetectionEnabled=%s, applicationNumber=%d",
+  return [NSString stringWithFormat:@"<%s %p - method=%@, uri=%@, httpVersion=%@, headers=%@, content=%@, userInfo=%@, defaultFormValueEncoding=%lu, formValueEncoding=%lu, formValues=%@, uriElements=%@, cookie=%@, applicationURLPrefix=%@, requestHandlerPathArray=%@, browserLanguages=%@, requestType=%d, isUsingWebServer=%s, formValueEncodingDetectionEnabled=%s, applicationNumber=%d",
                    object_getClassName(self),
                    (void*)self,
                    _method,
@@ -617,8 +617,8 @@ RCS_ID("$Id$")
                    _headers,
                    _contentData,
                    _userInfo,
-                   _defaultFormValueEncoding,
-                   _formValueEncoding,
+                   (unsigned long)_defaultFormValueEncoding,
+                   (unsigned long)_formValueEncoding,
                    _formValues,
                    _uriElements,
                    _cookie,
